+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/roadrunner/opmode/DriveVelocityPIDTuner.java
@@ -20,7 +20,6 @@ import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.RobotLog;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.teamcode.subsystems.drivetrains.MecanumDrivetrain;
 
 import java.util.List;
