+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/roadrunner/opmode/ManualFeedforwardTuner.java
@@ -22,7 +22,6 @@ import com.qualcomm.robotcore.hardware.VoltageSensor;
 import com.qualcomm.robotcore.util.RobotLog;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.teamcode.subsystems.drivetrains.MecanumDrivetrain;
 
 import java.util.Objects;
