+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/roadrunner/opmode/MotorDirectionDebugger.java
@@ -8,7 +8,6 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.teamcode.subsystems.drivetrains.MecanumDrivetrain;
 
 /**
  * This is a simple teleop routine for debugging your motor configuration.
