+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/roadrunner/opmode/StraightTest.java
@@ -9,7 +9,6 @@ import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.teamcode.subsystems.drivetrains.MecanumDrivetrain;
 
 /*
  * This is a simple routine to test translational drive capabilities.
