+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/roadrunner/opmode/TrackWidthTuner.java
@@ -12,7 +12,6 @@ import com.qualcomm.robotcore.util.MovingStatistics;
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.robotcore.internal.system.Misc;
 import org.firstinspires.ftc.teamcode.roadrunner.DriveConstants;
-import org.firstinspires.ftc.teamcode.subsystems.drivetrains.MecanumDrivetrain;
 
 /*
  * This routine determines the effective track width. The procedure works by executing a point turn
